Relocation of Messaging System
