+++ vcell-apiclient/src/main/java/org/vcell/api/messaging/LocalDataSetControllerMessaging.java
@@ -8,38 +8,32 @@
  *  http://www.opensource.org/licenses/mit-license.php
  */
 
-package cbit.vcell.message.server.bootstrap.client;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.vcell.api.client.VCellApiClient;
-import org.vcell.restclient.model.*;
-import org.vcell.solver.nfsim.NFSimMolecularConfigurations;
-import org.vcell.util.DataAccessException;
-import org.vcell.util.document.ExternalDataIdentifier;
-import org.vcell.util.document.KeyValue;
-import org.vcell.util.document.UserLoginInfo;
-import org.vcell.util.document.VCDataIdentifier;
-import org.vcell.vis.io.VtuFileContainer;
-import org.vcell.vis.io.VtuVarInfo;
+package org.vcell.api.messaging;
 
 import cbit.plot.PlotData;
 import cbit.rmi.event.ExportEvent;
 import cbit.vcell.export.server.ExportSpecs;
 import cbit.vcell.field.io.FieldDataFileOperationResults;
 import cbit.vcell.field.io.FieldDataFileOperationSpec;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
+import cbit.vcell.message.server.bootstrap.client.RpcDataServerProxy;
+import cbit.vcell.message.server.bootstrap.client.RpcSender;
 import cbit.vcell.server.DataSetController;
-import cbit.vcell.simdata.DataIdentifier;
-import cbit.vcell.simdata.DataOperation;
-import cbit.vcell.simdata.DataOperationResults;
-import cbit.vcell.simdata.DataSetMetadata;
-import cbit.vcell.simdata.DataSetTimeSeries;
-import cbit.vcell.simdata.OutputContext;
-import cbit.vcell.simdata.ParticleDataBlock;
-import cbit.vcell.simdata.SimDataBlock;
-import cbit.vcell.simdata.SpatialSelection;
+import cbit.vcell.simdata.*;
 import cbit.vcell.solvers.CartesianMesh;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.vcell.api.client.VCellApiClient;
+import org.vcell.restclient.model.AnalyzedResultsFromFieldData;
+import org.vcell.restclient.model.FieldDataSaveResults;
+import org.vcell.restclient.model.FieldDataShape;
+import org.vcell.restclient.utils.DtoModelTransforms;
+import org.vcell.solver.nfsim.NFSimMolecularConfigurations;
+import org.vcell.util.DataAccessException;
+import org.vcell.util.document.UserLoginInfo;
+import org.vcell.util.document.VCDataIdentifier;
+import org.vcell.vis.io.VtuFileContainer;
+import org.vcell.vis.io.VtuVarInfo;
 
 /**
  * This interface was generated by a SmartGuide.
@@ -64,15 +58,15 @@ public FieldDataFileOperationResults fieldDataFileOperation(FieldDataFileOperati
 	if (lg.isTraceEnabled()) lg.trace("LocalDataSetControllerMessaging.fieldDataFileOperationSpec(...)");
 	try {
 		if (fieldDataFileOperationSpec.opType == FieldDataFileOperationSpec.FDOS_ADD){
-			AnalyzedResultsFromFieldData analyzedResultsFromFieldData = FieldDataFileOperationSpec.fieldDataSpecToAnalyzedResultsDTO(fieldDataFileOperationSpec);
+			AnalyzedResultsFromFieldData analyzedResultsFromFieldData = DtoModelTransforms.fieldDataSpecToAnalyzedResultsDTO(fieldDataFileOperationSpec);
 			FieldDataSaveResults results = vCellApiClient.getFieldDataApi().createFieldDataFromAnalyzedFile(analyzedResultsFromFieldData);
-			return FieldDataFileOperationResults.fieldDataSaveResultsDTOToFileOperationResults(results, fieldDataFileOperationSpec.owner);
+			return DtoModelTransforms.fieldDataSaveResultsDTOToFileOperationResults(results, fieldDataFileOperationSpec.owner);
 		} else if (fieldDataFileOperationSpec.opType == FieldDataFileOperationSpec.FDOS_DELETE) {
 			vCellApiClient.getFieldDataApi().deleteFieldData(fieldDataFileOperationSpec.specEDI.getKey().toString());
 			return null;
 		} else if (fieldDataFileOperationSpec.opType == FieldDataFileOperationSpec.FDOS_INFO) {
 			FieldDataShape fieldDataInfo = vCellApiClient.getFieldDataApi().getFieldDataShapeFromID(fieldDataFileOperationSpec.sourceSimDataKey.toString());
-			return FieldDataFileOperationResults.fieldDataInfoDTOToFileOperationResults(fieldDataInfo);
+			return DtoModelTransforms.fieldDataInfoDTOToFileOperationResults(fieldDataInfo);
 		} else {
 			return dataServerProxy.fieldDataFileOperation(fieldDataFileOperationSpec);
 		}
@@ -89,7 +83,7 @@ public FieldDataFileOperationResults fieldDataFileOperation(FieldDataFileOperati
 /**
  * This method was created by a SmartGuide.
  * @return java.lang.String[]
- * @throws RemoteProxyException 
+ * @throws RemoteProxyException
  */
 public DataIdentifier[] getDataIdentifiers(OutputContext outputContext,VCDataIdentifier vcdID) throws DataAccessException, RemoteProxyException {
 	if (lg.isTraceEnabled()) lg.trace("LocalDataSetControllerMessaging.getDataIdentifiers(vcdID=" + vcdID + ")");
