+++ vcell-apiclient/src/main/java/org/vcell/api/messaging/LocalVCellConnectionMessaging.java
@@ -8,25 +8,25 @@
  *  http://www.opensource.org/licenses/mit-license.php
  */
 
-package cbit.vcell.message.server.bootstrap.client;
-import java.io.IOException;
+package org.vcell.api.messaging;
 
+import cbit.rmi.event.MessageEvent;
+import cbit.vcell.message.server.bootstrap.client.LocalSimulationControllerMessaging;
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
+import cbit.vcell.message.server.bootstrap.client.RpcSender;
+import cbit.vcell.mongodb.VCMongoMessage;
+import cbit.vcell.resource.ErrorUtils;
+import cbit.vcell.server.DataSetController;
+import cbit.vcell.server.SimulationController;
+import cbit.vcell.server.UserMetaDbServer;
+import cbit.vcell.server.VCellConnection;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.vcell.api.client.VCellApiClient;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.UserLoginInfo;
 
-import cbit.rmi.event.MessageEvent;
-import cbit.rmi.event.PerformanceMonitorEvent;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
-import cbit.vcell.mongodb.VCMongoMessage;
-import cbit.vcell.resource.ErrorUtils;
-import cbit.vcell.server.DataSetController;
-import cbit.vcell.server.PerformanceMonitoringFacility;
-import cbit.vcell.server.SimulationController;
-import cbit.vcell.server.UserMetaDbServer;
-import cbit.vcell.server.VCellConnection;
+import java.io.IOException;
 
 /**
  * The user's connection to the Virtual Cell.  It is obtained from the VCellServer
