+++ vcell-apiclient/src/test/java/org/vcell/api/messaging/RemoteProxyVCellConnectionFactoryTest.java
@@ -1,6 +1,6 @@
-package cbit.vcell.message.server.bootstrap.client;
+package org.vcell.api.messaging;
 
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import cbit.vcell.server.VCellConnection;
 import org.junit.jupiter.api.*;
 import org.vcell.api.client.VCellApiClient;
@@ -8,7 +8,6 @@ import org.vcell.api.client.query.BioModelsQuerySpec;
 import org.vcell.api.common.BiomodelRepresentation;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.UserLoginInfo;
-import org.vcell.util.document.UserLoginInfo.DigestedPassword;
 import org.vcell.util.document.VCInfoContainer;
 
 import java.io.IOException;
